Remove coupling from Shape to Entity and make shapes immutable
Cleanup to remove coupling from Shape to Entity and also made the shapes
immutable.  This also allows a shape to be used with more than 1 entity
at a time.

I rushed and was somewhat lazy so I just passed the entity references
when needed.  This could be made cleaner when we have a closer look at
it.