Improve code re-use by separating behaviors by using default methods in Interfaces
This is an idea for how we can improve code re-use.

One of the things that ECS systems fixes is that you end up with
multiple classes that have to re-implement common code because they
cannot share the same parents in the hierarchy.  Java 8 introduced
default methods in Interfaces so we can take advantage of that.

When defining a new class, we would simply define the new class as
implementing all the interfaces with the default behaviors that we want
(without needing to re-implement those behaviors).  So if a class
implements DrawShape then its shape is automatically drawn.